+++ additionalResolvers/content.ts
@@ -1,5 +1,49 @@
 import {Resolvers} from "../.mesh";
 
+/**
+ * Helper function to get a list of content ids the user has access to.
+ * @param context GraphQL context
+ * @param root  GraphQL root
+ * @param courseWhitelist Additional whitelist of course ids. If not null, only contents which the user has access
+ * to AND are in at least one of the courses in the whitelist are returned
+ * @returns List of content ids the user has access to
+ */
+async function getContentIdsUserHasAccessTo(context, root, courseWhitelist) {
+    // get a list of all courses the user has access to
+    let courseMembershipsRes = await context.CourseService.Query._internal_noauth_courseMembershipsByUserId({
+        root,
+        args: {
+            userId: context.currentUser.id,
+            availabilityFilter: true
+        },
+        selectionSet: `
+        {
+            courseId
+        }
+        `
+    });
+
+    // get a list of all media records in the courses the user has access to and, in case a whitelist is 
+    // provided, only consider the courses in the whitelist
+    let mediaRecordsRes = await context.ContentService.Query._internal_noauth_contentsAvailableToBeWorkedOnByUserForCourses({
+        root,
+        args: {
+            courseIds: courseMembershipsRes
+                .map((membership) => membership.courseId)
+                .filter((courseId) => courseWhitelist?.includes(courseId) ?? true)
+        },
+        selectionSet: `
+        {
+            id
+        }
+        `
+    });
+
+    let contentWhitelist = mediaRecordsRes.flat().map((mediaRecord) => mediaRecord.id);
+
+    return contentWhitelist;
+}
+
 const resolvers: Resolvers = {
     Mutation: {
         createMediaContentAndLinkRecords: {
@@ -189,6 +233,43 @@ const resolvers: Resolvers = {
         },
 
 
+    },
+    Query: {
+        semanticSearch: {
+            async resolve(root, args, context, info) {
+                const contentWhitelist = await getContentIdsUserHasAccessTo(context, root, args.courseWhitelist);
+
+                // run the semantic search
+                return context.DocprocaiService.Query._internal_noauth_semanticSearch({
+                    root,
+                    args: {
+                        queryText: args.queryText,
+                        count: args.count,
+                        contentWhitelist: contentWhitelist
+                    },
+                    context,
+                    info
+                });
+            }
+        },
+        getSemanticallySimilarEntities: {
+            async resolve(root, args, context, info) {
+                let contentWhitelist: string[] = await getContentIdsUserHasAccessTo(context, root, null);
+    
+                // get semantically similar segments
+                return context.DocprocaiService.Query._internal_noauth_getSemanticallySimilarMediaRecordSegments({
+                    root,
+                    args: {
+                        segmentId: args.mediaRecordSegmentId,
+                        count: args.count,
+                        contentWhitelist: contentWhitelist,
+                        excludeEntitiesWithSameParent: args.excludeEntitiesWithSameParent
+                    },
+                    context,
+                    info
+                });
+            }
+        }
     }
 }
